requirements.txt
beautifulsoup4
requests
httpx
python-dotenv
groq
flask
flask-login 
pymongo 
flask-bcrypt  
langchain
flask-socketio


streamlit
langchain-groq
langchain-community
langchain
langchain-core
faiss-cpu
langchain-community
